Updated Product Entity: Added price column. Improvements on model.

+++ src/main/java/com/bo/acredito/domain/Occupation.java
@@ -6,20 +6,20 @@ import javax.persistence.Entity;
 import javax.persistence.Id;
 
 /**
- * Created by aralco on 2/1/14.
+ * Created by aralco on 2/4/14.
  */
 @Entity
 public class Occupation {
-    private Integer id;
+    private Long id;
     private String name;
 
     @Id
     @Column(name = "id")
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
